safari-600.7-branch/LayoutTests/fast/forms/validationMessage-expected.txt

Test for validationMessage DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


input patternMismatch: pattern mismatch
input valueMissing: value missing
textarea valueMissing: value missing
select valueMissing: value missing
input typeMismatch: type mismatch
input badInput: type mismatch
badInput and valueMissing:
PASS numberInput.validationMessage is nonRequiredBadInputMessage
PASS but.validationMessage is ''
PASS anoninput.validationMessage is ''
PASS happyFieldset.validationMessage is ''
PASS happySelect.validationMessage is ''
PASS happyOutput.validationMessage is ''
PASS happyObject.validationMessage is ''
PASS happyKeygen.validationMessage is ''
PASS successfullyParsed is true

TEST COMPLETE
